Introduce resource tree snapshot feature to the Query API
This feature provides a "TakeSnapshot" API function which traverses all
or part of the Data Hub resource tree and encodes the results using
formatter plugins.  Various parameters control which information is
returned and how it is formatted.

The full implementation is a pair of interlocked state machines, one in
the snapshot core and one in the formatter plugin.  The snapshot code
walks through the resource tree in a depth-first traversal, and for
each node the formatter generates output and writes it to the output
stream file descriptor.  All of the steps take advantage of the Legato
event loop to maintain sequencing and to allow flow control on the
output stream.

This initial commit introduces the snapshot core code, but does not
include any formatters.  It also does not yet support deletion tracking
of resource tree nodes.

Resolves: LE-13032
Change-Id: I2ec9993b942c9f37c98dfbcaf7e8aa1c6343b8ad
(cherry picked from commit b326315161ebd3014ec4a730bb123a8971ef53a4)
Signed-off-by: imorrison <imorrison@sierrawireless.com>