BROOKLYN-3546 - Merge snapshot API from legato datahub (#17)
* Introduce resource tree snapshot feature to the Query API

This feature provides a "TakeSnapshot" API function which traverses all
or part of the Data Hub resource tree and encodes the results using
formatter plugins.  Various parameters control which information is
returned and how it is formatted.

The full implementation is a pair of interlocked state machines, one in
the snapshot core and one in the formatter plugin.  The snapshot code
walks through the resource tree in a depth-first traversal, and for
each node the formatter generates output and writes it to the output
stream file descriptor.  All of the steps take advantage of the Legato
event loop to maintain sequencing and to allow flow control on the
output stream.

This initial commit introduces the snapshot core code, but does not
include any formatters.  It also does not yet support deletion tracking
of resource tree nodes.

Resolves: LE-13032
Change-Id: I2ec9993b942c9f37c98dfbcaf7e8aa1c6343b8ad
(cherry picked from commit b326315161ebd3014ec4a730bb123a8971ef53a4)
Signed-off-by: imorrison <imorrison@sierrawireless.com>

* Add JSON snapshot formatter plugin

The JSON formatter encodes the requested resource tree information as
JSON and streams it back to the snapshot requester.

Resolves: LE-13032
Change-Id: Ie9a714923951894b3febab3b1e711117a4a315a1
(cherry picked from commit 7eeae55683769e202e1885a369e6c03f3c7bb82b)
Signed-off-by: imorrison <imorrison@sierrawireless.com>

* Add Data Hub snapshot tool

This tool provides a command line interface for invoking the snapshot
API and outputting the results to stdout or to a file.

Resolves: LE-13032
Change-Id: Icbe98dfcd39b26308fb9cfca0b74cfc85afd1842
(cherry picked from commit 5643070ef59478c1b352edbd052114625dc985b0)
Signed-off-by: imorrison <imorrison@sierrawireless.com>

* Implement tracking of deleted items in snapshot API

Now snapshot requests can include information on resource tree nodes
deleted since the last flush of the tree.  As part of this, formatters
can request multiple sequential scans of the resource tree filtered by
node state as part of a snapshot operation.  As an example, this allows
the existing nodes to be considered in one pass, and the deleted nodes
to be considered in another.

Each node now has three states indicating its lifetime:
 * "New," indicating it has never been captured by a previous snapshot.
 * "Deleted," indicating the node has been deleted.  Deleted nodes are
   only provided if deletion tracking is turned on.
 * "Normal," indicating the node is in the regular portion of its
   existence.

A pair of periodically added and removed resources have been added to
the actuator demo program to assist with testing.

Resolves: LE-14968
Change-Id: I979ad3d83ae3000d4c820a6948ae2a651e4bbb18
(cherry picked from commit 945cb9bec451d4ddd400af6f5b1a5fb217d80442)
Signed-off-by: imorrison <imorrison@sierrawireless.com>

* Update snapshot API for Octave compatibility

 - Add support for Octave CBOR-based format to snapshot API (octaveFormatter)
 - Only clear node's newness flag at the end of snapshot (add a resource flag)
 - Track resource's JSON example modification (add a resource flag)
 - Make some resource tree function public for Octave formatter use
 - Do not close read end of pipe from dataHub
 - Minor improvements in error handling

Resolves: LE-14969
Change-Id: Ibee72c8c64ce760cfb09998d87c1cea2038062dc

* Update snapshot API for Octave compatibility

 - Add support for Octave CBOR-based format to snapshot API (octaveFormatter)
 - Only clear node's newness flag at the end of snapshot (add a resource flag)
 - Track resource's JSON example modification (add a resource flag)
 - Make some resource tree function public for Octave formatter use
 - Do not close read end of pipe from dataHub
 - Minor improvements in error handling

Resolves: LE-14969
Change-Id: Ibee72c8c64ce760cfb09998d87c1cea2038062dc

Co-authored-by: Andrew MacIsaac <amacisaac@sierrawireless.com>
Co-authored-by: Yohan Marchiset <ymarchiset@sierrawireless.com>
Co-authored-by: imorrison <imorrison@sierrawireless.com>